Tratamento de Exceções Global Persistência em JSON Tratamento de Exceções Específico, paciente

+++ src/main/java/com/sistemaubs/gestao/service/PacienteService.java
@@ -1,11 +1,13 @@
 package com.sistemaubs.gestao.service;
 
 
-import com.sistemaubs.gestao.Repository.PacienteRepository;
+import com.sistemaubs.gestao.repository.PacienteRepository;
 import com.sistemaubs.gestao.model.Paciente;
+import com.sistemaubs.gestao.exception.PacienteNaoEncontradoException;
 
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class PacienteService {
@@ -26,16 +28,30 @@ public class PacienteService {
     }
 
     public void removerPaciente(Long id) {
+        Paciente paciente = pacienteRepository.pegarPacienteId(id);
+
+        if(paciente == null){
+            throw new PacienteNaoEncontradoException("Paciente com ID " + id + " não encontrado");
+        }
 
         pacienteRepository.removerPaciente(id);
     }
 
     public Paciente pegarPacienteId(Long id){
+        Paciente paciente = pacienteRepository.pegarPacienteId(id);
+
+        if (paciente == null) {
+            throw new PacienteNaoEncontradoException("Paciente com ID " + id + " não encontrado");
+        }
         return pacienteRepository.pegarPacienteId(id);
 
     }
 
     public Paciente editarPaciente(Long id, Paciente pacienteEditado) {
+        Paciente paciente = pacienteRepository.pegarPacienteId(id);
+        if (paciente == null) {
+            throw new PacienteNaoEncontradoException("Paciente com ID " + id + " não encontrado");
+        }
         return pacienteRepository.editarPaciente(id, pacienteEditado);
     }
 }
